+++ packages/magento2-module-inpost-pay/src/Model/IziApi/Request/UpdateOrderRequest.php
@@ -20,6 +20,8 @@ class UpdateOrderRequest extends Request implements RequestInterface
 
     protected ?string $contentType = 'application/json';
 
+    private ?int $storeId = null;
+
     /**
      * @param IziApiConfigProvider $iziApiConfigProvider
      * @param TokenGenerator $tokenGenerator
@@ -50,11 +52,24 @@ class UpdateOrderRequest extends Request implements RequestInterface
 
     public function getApiUrl(): string
     {
-        return $this->iziApiConfigProvider->getIziApiUrl();
+        $storeId = $this->storeId ?? 0;
+
+        return $this->iziApiConfigProvider->getIziApiUrl($storeId);
     }
 
     public function getBearerToken(): ?string
     {
-        return $this->tokenGenerator->generate()->getAccessToken();
+        $storeId = $this->storeId ?? 0;
+
+        return $this->tokenGenerator->generate(false, $storeId)->getAccessToken();
+    }
+
+    /**
+     * @param int $storeId
+     * @return void
+     */
+    public function setStoreId(int $storeId): void
+    {
+        $this->storeId = $storeId;
     }
 }