+++ packages/magento2-module-inpost-pay/src/Provider/Config/GeneralConfigProvider.php
@@ -14,6 +14,7 @@ class GeneralConfigProvider
     private const XML_PATH_INPOST_PAY_NEW_ORDER_STATUS = 'payment/inpost_pay/order_status';
     private const XML_PATH_ORDER_ADDRESS_SOURCE_FLAG = 'payment/inpost_pay/use_address_as_firstname_source';
     private const XML_PATH_INPOST_PAY_IMAGE_ROLE = 'payment/inpost_pay/image_role';
+    private const XML_PATH_INPOST_PAY_ASSIGN_QUOTE_ENABLED = 'payment/inpost_pay/assign_quote_to_customer_by_email';
 
     /**
      * @param ScopeConfigInterface $scopeConfig
@@ -46,13 +47,15 @@ class GeneralConfigProvider
     }
 
     /**
+     * @param int|null $storeId
      * @return string
      */
-    public function getNewOrderStatus(): string
+    public function getNewOrderStatus(?int $storeId = null): string
     {
         $orderStatus = $this->scopeConfig->getValue(
             self::XML_PATH_INPOST_PAY_NEW_ORDER_STATUS,
-            ScopeInterface::SCOPE_WEBSITE
+            ScopeInterface::SCOPE_STORE,
+            $storeId
         );
 
         return is_scalar($orderStatus) ? (string)$orderStatus : 'pending';
@@ -71,5 +74,14 @@ class GeneralConfigProvider
         );
 
         return is_scalar($orderStatus) ? (string)$orderStatus : 'small_image';
+    }
+
+    public function isAssigningGuestCartsToAccountByEmailEnabled(?int $storeId = null): bool
+    {
+        return $this->scopeConfig->isSetFlag(
+            self::XML_PATH_INPOST_PAY_ASSIGN_QUOTE_ENABLED,
+            ScopeInterface::SCOPE_STORE,
+            $storeId
+        );
     }
 }