+++ packages/magento2-module-inpost-pay/src/Service/Order/Creator/Steps/AssignCustomerStep.php
@@ -6,6 +6,7 @@ namespace InPost\InPostPay\Service\Order\Creator\Steps;
 
 use InPost\InPostPay\Api\OrderProcessingStepInterface;
 use InPost\InPostPay\Api\Data\Merchant\OrderInterface;
+use InPost\InPostPay\Provider\Config\GeneralConfigProvider;
 use Magento\Customer\Api\CustomerRepositoryInterface;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NoSuchEntityException;
@@ -16,6 +17,7 @@ class AssignCustomerStep extends OrderProcessingStep implements OrderProcessingS
 {
     public function __construct(
         private readonly CustomerRepositoryInterface $customerRepository,
+        private readonly GeneralConfigProvider $generalConfigProvider,
         LoggerInterface $logger
     ) {
         parent::__construct($logger);
@@ -24,25 +26,35 @@ class AssignCustomerStep extends OrderProcessingStep implements OrderProcessingS
     public function process(Quote $quote, OrderInterface $inPostOrder): void
     {
         // @phpstan-ignore-next-line
-        $quoteCustomerId = (int)$quote->getCustomer()->getId();
-        $email = $inPostOrder->getAccountInfo()->getMail();
+        $quoteCustomerId = is_scalar($quote->getCustomer()->getId()) ? (int)$quote->getCustomer()->getId() : null;
+        $accountEmail = $inPostOrder->getAccountInfo()->getMail();
         $websiteId = (int)$quote->getStore()->getWebsiteId();
+
         try {
-            $customer = $this->customerRepository->get($email, $websiteId);
-            $foundCustomerId = (int)$customer->getId();
-            if ($foundCustomerId && $quoteCustomerId === $foundCustomerId) {
+            if ($quoteCustomerId) {
+                $this->createLog(
+                    sprintf(
+                        'Basket has been initialized by logged-in user. Order will be assigned to customer ID: %s',
+                        (int)$quote->getCustomerId()
+                    )
+                );
+            } elseif ($this->generalConfigProvider->isAssigningGuestCartsToAccountByEmailEnabled($quote->getStoreId())
+                && $accountEmail
+            ) {
+                $customer = $this->customerRepository->get($accountEmail, $websiteId);
                 $quote->assignCustomer($customer);
+                $quote->setCustomerIsGuest(false);
                 $this->createLog(
                     sprintf(
                         'Customer account found by email: %s. Order will be assigned to customer ID: %s',
-                        $email,
+                        $accountEmail,
                         (int)$customer->getId()
                     )
                 );
             }
         } catch (NoSuchEntityException | LocalizedException $e) {
             $this->createLog(
-                sprintf('Customer account not found by email: %s. Order will be processed for guest.', $email)
+                sprintf('Customer account not found by email: %s. Order will be processed for guest.', $accountEmail)
             );
         }
     }