+++ packages/magento2-module-inpost-pay/src/view/frontend/web/js/inpostPayMerchantWidget.js
@@ -101,7 +101,6 @@ define([
              * @property {string} language
              * @property {unboundWidgetClicked} unboundWidgetClicked
              * @property {handleBasketEvent} handleBasketEvent
-             * @property {string} apiBaseUrl
              * @property {boolean} webView
              */
             var widgetOptions = $.extend({
@@ -111,7 +110,6 @@ define([
                 handleBasketEvent: this.handleBasketEvent.bind(this),
             }, {
                 language: config.language ? config.language : undefined,
-                apiBaseUrl: config.apiBaseUrl ? config.apiBaseUrl : undefined,
                 webView: config.webView ? config.webView : undefined,
             });